[Tiered Caching] Gate CacheStatsHolder logic behind FeatureFlags.PLUGGABLE_CACHE setting (#13238)
Stats rework step 2 of 4

---------

(cherry picked from commit f84d28d4371dbd00356cccc01366d87d21814f36)
