+++ server/src/main/java/org/opensearch/indices/IndicesRequestCache.java
@@ -801,13 +801,6 @@ public final class IndicesRequestCache implements RemovalListener<ICacheKey<Indi
     }
 
     /**
-     * Returns the current size in bytes of the cache
-     */
-    long getSizeInBytes() {
-        return cache.stats().getTotalSizeInBytes();
-    }
-
-    /**
      * Returns the current cache stats. Pkg-private for testing.
      */
     ImmutableCacheStatsHolder stats() {
