+++ server/src/test/java/org/opensearch/common/cache/stats/DefaultCacheStatsHolderTests.java
@@ -21,18 +21,23 @@ import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CountDownLatch;
 
-public class CacheStatsHolderTests extends OpenSearchTestCase {
+public class DefaultCacheStatsHolderTests extends OpenSearchTestCase {
     public void testAddAndGet() throws Exception {
         List<String> dimensionNames = List.of("dim1", "dim2", "dim3", "dim4");
-        CacheStatsHolder cacheStatsHolder = new CacheStatsHolder(dimensionNames);
-        Map<String, List<String>> usedDimensionValues = CacheStatsHolderTests.getUsedDimensionValues(cacheStatsHolder, 10);
-        Map<List<String>, CacheStats> expected = CacheStatsHolderTests.populateStats(cacheStatsHolder, usedDimensionValues, 1000, 10);
+        DefaultCacheStatsHolder cacheStatsHolder = new DefaultCacheStatsHolder(dimensionNames);
+        Map<String, List<String>> usedDimensionValues = DefaultCacheStatsHolderTests.getUsedDimensionValues(cacheStatsHolder, 10);
+        Map<List<String>, CacheStats> expected = DefaultCacheStatsHolderTests.populateStats(
+            cacheStatsHolder,
+            usedDimensionValues,
+            1000,
+            10
+        );
 
         // test the value in the map is as expected for each distinct combination of values
         for (List<String> dimensionValues : expected.keySet()) {
             CacheStats expectedCounter = expected.get(dimensionValues);
 
-            ImmutableCacheStats actualStatsHolder = CacheStatsHolderTests.getNode(dimensionValues, cacheStatsHolder.getStatsRoot())
+            ImmutableCacheStats actualStatsHolder = DefaultCacheStatsHolderTests.getNode(dimensionValues, cacheStatsHolder.getStatsRoot())
                 .getImmutableStats();
             ImmutableCacheStats actualCacheStats = getNode(dimensionValues, cacheStatsHolder.getStatsRoot()).getImmutableStats();
 
@@ -53,7 +58,7 @@ public class CacheStatsHolderTests extends OpenSearchTestCase {
 
     public void testReset() throws Exception {
         List<String> dimensionNames = List.of("dim1", "dim2");
-        CacheStatsHolder cacheStatsHolder = new CacheStatsHolder(dimensionNames);
+        DefaultCacheStatsHolder cacheStatsHolder = new DefaultCacheStatsHolder(dimensionNames);
         Map<String, List<String>> usedDimensionValues = getUsedDimensionValues(cacheStatsHolder, 10);
         Map<List<String>, CacheStats> expected = populateStats(cacheStatsHolder, usedDimensionValues, 100, 10);
 
@@ -64,7 +69,7 @@ public class CacheStatsHolderTests extends OpenSearchTestCase {
             originalCounter.sizeInBytes = new CounterMetric();
             originalCounter.entries = new CounterMetric();
 
-            CacheStatsHolder.Node node = getNode(dimensionValues, cacheStatsHolder.getStatsRoot());
+            DefaultCacheStatsHolder.Node node = getNode(dimensionValues, cacheStatsHolder.getStatsRoot());
             ImmutableCacheStats actual = node.getImmutableStats();
             assertEquals(originalCounter.immutableSnapshot(), actual);
         }
@@ -72,7 +77,7 @@ public class CacheStatsHolderTests extends OpenSearchTestCase {
 
     public void testDropStatsForDimensions() throws Exception {
         List<String> dimensionNames = List.of("dim1", "dim2");
-        CacheStatsHolder cacheStatsHolder = new CacheStatsHolder(dimensionNames);
+        DefaultCacheStatsHolder cacheStatsHolder = new DefaultCacheStatsHolder(dimensionNames);
 
         // Create stats for the following dimension sets
         List<List<String>> populatedStats = List.of(List.of("A1", "B1"), List.of("A2", "B2"), List.of("A2", "B3"));
@@ -108,7 +113,7 @@ public class CacheStatsHolderTests extends OpenSearchTestCase {
 
     public void testCount() throws Exception {
         List<String> dimensionNames = List.of("dim1", "dim2");
-        CacheStatsHolder cacheStatsHolder = new CacheStatsHolder(dimensionNames);
+        DefaultCacheStatsHolder cacheStatsHolder = new DefaultCacheStatsHolder(dimensionNames);
         Map<String, List<String>> usedDimensionValues = getUsedDimensionValues(cacheStatsHolder, 10);
         Map<List<String>, CacheStats> expected = populateStats(cacheStatsHolder, usedDimensionValues, 100, 10);
 
@@ -121,7 +126,7 @@ public class CacheStatsHolderTests extends OpenSearchTestCase {
 
     public void testConcurrentRemoval() throws Exception {
         List<String> dimensionNames = List.of("dim1", "dim2");
-        CacheStatsHolder cacheStatsHolder = new CacheStatsHolder(dimensionNames);
+        DefaultCacheStatsHolder cacheStatsHolder = new DefaultCacheStatsHolder(dimensionNames);
 
         // Create stats for the following dimension sets
         List<List<String>> populatedStats = List.of(List.of("A1", "B1"), List.of("A2", "B2"), List.of("A2", "B3"));
@@ -169,8 +174,8 @@ public class CacheStatsHolderTests extends OpenSearchTestCase {
      * Returns the node found by following these dimension values down from the root node.
      * Returns null if no such node exists.
      */
-    static CacheStatsHolder.Node getNode(List<String> dimensionValues, CacheStatsHolder.Node root) {
-        CacheStatsHolder.Node current = root;
+    static DefaultCacheStatsHolder.Node getNode(List<String> dimensionValues, DefaultCacheStatsHolder.Node root) {
+        DefaultCacheStatsHolder.Node current = root;
         for (String dimensionValue : dimensionValues) {
             current = current.getChildren().get(dimensionValue);
             if (current == null) {
@@ -181,7 +186,7 @@ public class CacheStatsHolderTests extends OpenSearchTestCase {
     }
 
     static Map<List<String>, CacheStats> populateStats(
-        CacheStatsHolder cacheStatsHolder,
+        DefaultCacheStatsHolder cacheStatsHolder,
         Map<String, List<String>> usedDimensionValues,
         int numDistinctValuePairs,
         int numRepetitionsPerValue
@@ -211,7 +216,7 @@ public class CacheStatsHolderTests extends OpenSearchTestCase {
                     expected.get(dimensions).evictions.inc(statsToInc.getEvictions());
                     expected.get(dimensions).sizeInBytes.inc(statsToInc.getSizeInBytes());
                     expected.get(dimensions).entries.inc(statsToInc.getEntries());
-                    CacheStatsHolderTests.populateStatsHolderFromStatsValueMap(cacheStatsHolder, Map.of(dimensions, statsToInc));
+                    DefaultCacheStatsHolderTests.populateStatsHolderFromStatsValueMap(cacheStatsHolder, Map.of(dimensions, statsToInc));
                 }
                 countDownLatch.countDown();
             });
@@ -240,7 +245,7 @@ public class CacheStatsHolderTests extends OpenSearchTestCase {
         return result;
     }
 
-    static Map<String, List<String>> getUsedDimensionValues(CacheStatsHolder cacheStatsHolder, int numValuesPerDim) {
+    static Map<String, List<String>> getUsedDimensionValues(DefaultCacheStatsHolder cacheStatsHolder, int numValuesPerDim) {
         Map<String, List<String>> usedDimensionValues = new HashMap<>();
         for (int i = 0; i < cacheStatsHolder.getDimensionNames().size(); i++) {
             List<String> values = new ArrayList<>();
@@ -252,20 +257,20 @@ public class CacheStatsHolderTests extends OpenSearchTestCase {
         return usedDimensionValues;
     }
 
-    private void assertSumOfChildrenStats(CacheStatsHolder.Node current) {
+    private void assertSumOfChildrenStats(DefaultCacheStatsHolder.Node current) {
         if (!current.children.isEmpty()) {
             CacheStats expectedTotal = new CacheStats();
-            for (CacheStatsHolder.Node child : current.children.values()) {
+            for (DefaultCacheStatsHolder.Node child : current.children.values()) {
                 expectedTotal.add(child.getImmutableStats());
             }
             assertEquals(expectedTotal.immutableSnapshot(), current.getImmutableStats());
-            for (CacheStatsHolder.Node child : current.children.values()) {
+            for (DefaultCacheStatsHolder.Node child : current.children.values()) {
                 assertSumOfChildrenStats(child);
             }
         }
     }
 
-    static void populateStatsHolderFromStatsValueMap(CacheStatsHolder cacheStatsHolder, Map<List<String>, CacheStats> statsMap) {
+    static void populateStatsHolderFromStatsValueMap(DefaultCacheStatsHolder cacheStatsHolder, Map<List<String>, CacheStats> statsMap) {
         for (Map.Entry<List<String>, CacheStats> entry : statsMap.entrySet()) {
             CacheStats stats = entry.getValue();
             List<String> dims = entry.getKey();
